refactor(services): add new folder for notifications for all notification service files
