+++ app/backend/src/services/notifications/severity.py
